test: wrap inject + code signing part into a helper
Refs: https://github.com/nodejs/node/pull/47588#discussion_r1182802436
